Move chart to helm package
