+++ cli/common/utils.go
@@ -8,7 +8,6 @@ import (
 
 	"helm.sh/helm/v3/pkg/action"
 	helmCLI "helm.sh/helm/v3/pkg/cli"
-	"k8s.io/cli-runtime/pkg/genericclioptions"
 )
 
 const (
@@ -29,20 +28,6 @@ const (
 func Abort(raw string) bool {
 	confirmation := strings.TrimSuffix(raw, "\n")
 	return !(strings.ToLower(confirmation) == "y" || strings.ToLower(confirmation) == "yes")
-}
-
-// InitActionConfig initializes a Helm Go SDK action configuration. This function currently uses a hack to override the
-// namespace field that gets set in the K8s client set up by the SDK.
-func InitActionConfig(actionConfig *action.Configuration, namespace string, settings *helmCLI.EnvSettings, logger action.DebugLog) (*action.Configuration, error) {
-	getter := settings.RESTClientGetter()
-	configFlags := getter.(*genericclioptions.ConfigFlags)
-	configFlags.Namespace = &namespace
-	err := actionConfig.Init(settings.RESTClientGetter(), namespace,
-		os.Getenv("HELM_DRIVER"), logger)
-	if err != nil {
-		return nil, fmt.Errorf("error setting up helm action configuration to find existing installations: %s", err)
-	}
-	return actionConfig, nil
 }
 
 // CheckForInstallations uses the helm Go SDK to find helm releases in all namespaces where the chart name is
