+++ cli/helm/action.go
@@ -1,0 +1,24 @@
+package helm
+
+import (
+	"fmt"
+	"os"
+
+	"helm.sh/helm/v3/pkg/action"
+	helmCLI "helm.sh/helm/v3/pkg/cli"
+	"k8s.io/cli-runtime/pkg/genericclioptions"
+)
+
+// InitActionConfig initializes a Helm Go SDK action configuration. This function currently uses a hack to override the
+// namespace field that gets set in the K8s client set up by the SDK.
+func InitActionConfig(actionConfig *action.Configuration, namespace string, settings *helmCLI.EnvSettings, logger action.DebugLog) (*action.Configuration, error) {
+	getter := settings.RESTClientGetter()
+	configFlags := getter.(*genericclioptions.ConfigFlags)
+	configFlags.Namespace = &namespace
+	err := actionConfig.Init(settings.RESTClientGetter(), namespace,
+		os.Getenv("HELM_DRIVER"), logger)
+	if err != nil {
+		return nil, fmt.Errorf("error setting up helm action configuration to find existing installations: %s", err)
+	}
+	return actionConfig, nil
+}
