+++ cli/helm/chart_test.go
@@ -1,4 +1,4 @@
-package common
+package helm
 
 import (
 	"embed"
