+++ cli/helm/fixtures/consul/templates/foo.yaml
